requirements.txt

arrow==0.17.0
boto3>=1.26.0
Brotli>=1.0.9
django>=4.2,<5.0
djangorestframework>=3.14.0
django-cas-ng>=5.0.0
django-configurations>=2.4.1
django-extensions>=3.2.0
django-fsm>=2.8.0
django-impersonate>=1.9.0
django-ipware>=7.0.1
django-phonenumber-field>=7.0.0
django-storages>=1.13.0
dj-database-url>=2.0.0
elasticsearch>=7.17.0,<8.0
dockerflow>=2022.8.0
gunicorn>=21.0.0
phonenumbers>=8.13.0
psycopg2-binary>=2.9.6
requests>=2.28.0
sentry-sdk>=1.14.0
whitenoise>=6.4.0
python-docx>=0.8.11
fpdf2>=2.6.1
mammoth>=1.5.0
pdfminer.six
python-slugify>=8.0.0